requirements.txt
acres==0.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
cssselect2==0.7.0
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.1.0
extra-streamlit-components==0.1.71
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.55.2
fpdf==1.7.2
frozenlist==1.5.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-text-splitters==0.3.2
langsmith==0.1.147
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.2
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.56.1
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.12
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20240706
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
prov==2.0.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyphen==0.15.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.39.0
streamlit-google-auth==1.1.8
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
weasyprint==53.3
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
